Move unused code to old.py and tests to test_old_code.py.
Add a second test_param file with non-zero eccentricity
