+++ tests/test_params2.txt
@@ -1,0 +1,28 @@
+# Some parameter description and units:
+# mean_val - Mean RV motion          (km/s)
+# k1 - semi-major axis of star       (km/s)
+# omega - Argument of periapsis       (deg)
+# tau - reference time                (JD*)
+# period - orbital period             (day)
+# m_star - Mass of host star        (M_sun)
+# r_star - Radius of star           (R_sun)
+# age - Age of star                   (Gyr)
+# a - Semimajor axis distance.         (AU)
+# msini - Minimum mass of companion (M Jup)
+# m_true - Companion true mass      (M_jup)
+# i - inclination                     (deg)
+# temp - Host star Simbad temperature   (K)
+# fe_h - Host star metalicity         (dex)
+# logg - Host star logg           (cm s^-2)
+# Source - List of sources
+name         = Test2
+mean_val     = -5
+k1           = 7
+omega        = 30
+eccentricity = 0.5
+tau          = 2454000
+period       = 10
+msini        = 10
+m_true       = 100
+m1           = 1.2
+k2           = -20  # for old test to pass
